:muscle: Change location of definition of `command`
