+++ denops/gin/command/bare/main.ts
@@ -1,8 +1,12 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.1.0/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
 import * as helper from "https://deno.land/x/denops_std@v4.1.0/helper/mod.ts";
-import { parseSilent } from "../../util/cmd.ts";
-import { command } from "./command.ts";
+import {
+  parseOpts,
+  validateOpts,
+} from "https://deno.land/x/denops_std@v4.1.0/argument/opts.ts";
+import { normCmdArgs, parseSilent } from "../../util/cmd.ts";
+import { exec } from "./command.ts";
 
 export function main(denops: Denops): void {
   denops.dispatcher = {
@@ -16,4 +20,20 @@ export function main(denops: Denops): void {
       });
     },
   };
+}
+
+async function command(denops: Denops, args: string[]): Promise<void> {
+  const [opts, residue] = parseOpts(await normCmdArgs(denops, args));
+  validateOpts(opts, [
+    "worktree",
+    "enc",
+    "encoding",
+    "ff",
+    "fileformat",
+  ]);
+  await exec(denops, residue, {
+    worktree: opts.worktree,
+    encoding: opts.enc ?? opts.encoding,
+    fileformat: opts.ff ?? opts.fileformat,
+  });
 }
