+++ denops/gin/command/branch/main.ts
@@ -1,8 +1,21 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.1.0/mod.ts";
 import * as helper from "https://deno.land/x/denops_std@v4.1.0/helper/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
-import { parseDisableDefaultArgs, parseSilent } from "../../util/cmd.ts";
-import { command } from "./command.ts";
+import * as vars from "https://deno.land/x/denops_std@v4.1.0/variable/mod.ts";
+import {
+  builtinOpts,
+  formatOpts,
+  parse,
+  validateFlags,
+  validateOpts,
+} from "https://deno.land/x/denops_std@v4.1.0/argument/mod.ts";
+
+import {
+  normCmdArgs,
+  parseDisableDefaultArgs,
+  parseSilent,
+} from "../../util/cmd.ts";
+import { exec } from "./command.ts";
 import { edit } from "./edit.ts";
 
 export function main(denops: Denops): void {
@@ -30,4 +43,51 @@ export function main(denops: Denops): void {
       return helper.friendlyCall(denops, () => edit(denops, bufnr, bufname));
     },
   };
+}
+
+const allowedFlags = [
+  "a",
+  "all",
+  "r",
+  "remotes",
+  "i",
+  "ignore-case",
+  "abbrev",
+  "no-abbrev",
+];
+
+type CommandOptions = {
+  disableDefaultArgs?: boolean;
+};
+
+async function command(
+  denops: Denops,
+  mods: string,
+  args: string[],
+  options: CommandOptions = {},
+): Promise<void> {
+  if (!options.disableDefaultArgs) {
+    const defaultArgs = await vars.g.get(
+      denops,
+      "gin_branch_default_args",
+      [],
+    );
+    unknownutil.assertArray(defaultArgs, unknownutil.isString);
+    args = [...defaultArgs, ...args];
+  }
+  const [opts, flags, residue] = parse(await normCmdArgs(denops, args));
+  validateOpts(opts, [
+    "worktree",
+    "opener",
+    ...builtinOpts,
+  ]);
+  validateFlags(flags, allowedFlags);
+  await exec(denops, {
+    worktree: opts.worktree,
+    patterns: residue,
+    flags,
+    opener: opts.opener,
+    cmdarg: formatOpts(opts, builtinOpts).join(" "),
+    mods,
+  });
 }
