+++ denops/gin/command/buffer/command.ts
@@ -5,37 +5,7 @@ import * as option from "https://deno.land/x/denops_std@v4.1.0/option/mod.ts";
 import {
   format as formatBufname,
 } from "https://deno.land/x/denops_std@v4.1.0/bufname/mod.ts";
-import {
-  builtinOpts,
-  formatOpts,
-  parseOpts,
-  validateOpts,
-} from "https://deno.land/x/denops_std@v4.1.0/argument/mod.ts";
-import { normCmdArgs } from "../../util/cmd.ts";
 import { findWorktreeFromDenops } from "../../git/worktree.ts";
-
-export async function command(
-  denops: Denops,
-  mods: string,
-  args: string[],
-): Promise<buffer.OpenResult> {
-  const [opts, residue] = parseOpts(await normCmdArgs(denops, args));
-  validateOpts(opts, [
-    "processor",
-    "worktree",
-    "monochrome",
-    "opener",
-    ...builtinOpts,
-  ]);
-  return exec(denops, residue, {
-    processor: opts.processor?.split(" "),
-    worktree: opts.worktree,
-    monochrome: unnullish(opts.monochrome, () => true),
-    opener: opts.opener,
-    cmdarg: formatOpts(opts, builtinOpts).join(" "),
-    mods,
-  });
-}
 
 export type ExecOptions = {
   processor?: string[];
