+++ denops/gin/command/patch/command.ts
@@ -6,54 +6,9 @@ import * as mapping from "https://deno.land/x/denops_std@v4.1.0/mapping/mod.ts";
 import * as vars from "https://deno.land/x/denops_std@v4.1.0/variable/mod.ts";
 import * as option from "https://deno.land/x/denops_std@v4.1.0/option/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
-import {
-  builtinOpts,
-  formatOpts,
-  parse,
-  validateOpts,
-} from "https://deno.land/x/denops_std@v4.1.0/argument/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.1.0/buffer/mod.ts";
-import { normCmdArgs } from "../../util/cmd.ts";
 import { findWorktreeFromDenops } from "../../git/worktree.ts";
 import { exec as execEdit } from "../edit/command.ts";
-
-export type CommandOptions = {
-  disableDefaultArgs?: boolean;
-};
-
-export async function command(
-  denops: Denops,
-  mods: string,
-  args: string[],
-  options: CommandOptions = {},
-): Promise<void> {
-  if (!options.disableDefaultArgs) {
-    const defaultArgs = await vars.g.get(
-      denops,
-      "gin_patch_default_args",
-      [],
-    );
-    unknownutil.assertArray(defaultArgs, unknownutil.isString);
-    args = [...defaultArgs, ...args];
-  }
-  const [opts, _, residue] = parse(await normCmdArgs(denops, args));
-  validateOpts(opts, [
-    "worktree",
-    "opener",
-    "no-head",
-    "no-worktree",
-    ...builtinOpts,
-  ]);
-  const [abspath] = parseResidue(residue);
-  await exec(denops, abspath, {
-    worktree: opts.worktree,
-    noHead: "no-head" in opts,
-    noWorktree: "no-worktree" in opts,
-    opener: opts.opener,
-    cmdarg: formatOpts(opts, builtinOpts).join(" "),
-    mods,
-  });
-}
 
 export type ExecOptions = {
   worktree?: string;
@@ -159,18 +114,6 @@ export async function exec(
 
   // Focus INDEX
   await fn.win_gotoid(denops, infoIndex.winid);
-}
-
-function parseResidue(
-  residue: string[],
-): [string] {
-  // GinPatch [{options}] {path}
-  switch (residue.length) {
-    case 1:
-      return [residue[0]];
-    default:
-      throw new Error("Invalid number of arguments");
-  }
 }
 
 async function initHead(
