+++ denops/gin/command/patch/main.ts
@@ -1,8 +1,19 @@
 import type { Denops } from "https://deno.land/x/denops_std@v4.1.0/mod.ts";
 import * as helper from "https://deno.land/x/denops_std@v4.1.0/helper/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
-import { parseDisableDefaultArgs, parseSilent } from "../../util/cmd.ts";
-import { command } from "./command.ts";
+import * as vars from "https://deno.land/x/denops_std@v4.1.0/variable/mod.ts";
+import {
+  builtinOpts,
+  formatOpts,
+  parse,
+  validateOpts,
+} from "https://deno.land/x/denops_std@v4.1.0/argument/mod.ts";
+import {
+  normCmdArgs,
+  parseDisableDefaultArgs,
+  parseSilent,
+} from "../../util/cmd.ts";
+import { exec } from "./command.ts";
 
 export function main(denops: Denops): void {
   denops.dispatcher = {
@@ -24,4 +35,54 @@ export function main(denops: Denops): void {
       });
     },
   };
+}
+
+type CommandOptions = {
+  disableDefaultArgs?: boolean;
+};
+
+async function command(
+  denops: Denops,
+  mods: string,
+  args: string[],
+  options: CommandOptions = {},
+): Promise<void> {
+  if (!options.disableDefaultArgs) {
+    const defaultArgs = await vars.g.get(
+      denops,
+      "gin_patch_default_args",
+      [],
+    );
+    unknownutil.assertArray(defaultArgs, unknownutil.isString);
+    args = [...defaultArgs, ...args];
+  }
+  const [opts, _, residue] = parse(await normCmdArgs(denops, args));
+  validateOpts(opts, [
+    "worktree",
+    "opener",
+    "no-head",
+    "no-worktree",
+    ...builtinOpts,
+  ]);
+  const [abspath] = parseResidue(residue);
+  await exec(denops, abspath, {
+    worktree: opts.worktree,
+    noHead: "no-head" in opts,
+    noWorktree: "no-worktree" in opts,
+    opener: opts.opener,
+    cmdarg: formatOpts(opts, builtinOpts).join(" "),
+    mods,
+  });
+}
+
+function parseResidue(
+  residue: string[],
+): [string] {
+  // GinPatch [{options}] {path}
+  switch (residue.length) {
+    case 1:
+      return [residue[0]];
+    default:
+      throw new Error("Invalid number of arguments");
+  }
 }
